pseudo codigo

+++ livro/capitulos/code/registros/reg_produto.c
@@ -1,5 +1,5 @@
 struct Produto {
-  int   codigo;
+  long  codigo;
   char  nome[100];
   float preco;
 };
